should be fully functioning+reworked

+++ thumbnail-generation/pipeline.py
@@ -1,14 +1,11 @@
-import argparse
 import sieve
-import concurrent
 import random
-import subprocess
 
 
 @sieve.function(name='video-thumbnails',
-                system_packages=["ffmpeg"],
-                run_commands=["git clone https://github.com/ashvash182/workflow-custom-fonts"])
-def workflow(video : sieve.Video, font : str, CLIP_prompts : str):
+                python_packages=["openai", "sentence_transformers", "Pillow"],
+                system_packages=["ffmpeg"])
+def main(video : sieve.Video, font : str, CLIP_prompts : str):
     '''
     :param video: A video input
     :param font: Desired font for video title (See https://github.com/ashvash182/workflow-custom-fonts)
@@ -21,6 +18,9 @@ def workflow(video : sieve.Video, font : str, CLIP_prompts : str):
     import openai
     from sentence_transformers import SentenceTransformer, util
     from PIL import Image
+
+    if not font:
+        font = "Bebas_Neue"
 
     print('loading CLIP...')
 
@@ -45,9 +45,9 @@ def workflow(video : sieve.Video, font : str, CLIP_prompts : str):
         # Output singular video frame, meaning no scenes extracted!
         return
                 
-    print('getting video title...')
+    # print('getting video title...')
 
-    video_title = list(sieve.function.get('ansh-sievedata-com/generate_video_title').run(video))
+    video_title = sieve.function.get('ansh-sievedata-com/generate_video_title').run(video)
 
     from sentence_transformers import SentenceTransformer, util
 
@@ -61,24 +61,22 @@ def workflow(video : sieve.Video, font : str, CLIP_prompts : str):
 
         hits = util.semantic_search(query_emb, img_emb, top_k=k)[0]
         
-        matched = []
         for hit in hits:
             image_path = img_names[hit['corpus_id']]  
-            matched.append(sieve.Image(path=image_path))
-            
-        CLIP_outputs.append(matched)
-        
+            CLIP_outputs.append(sieve.Image(path=image_path))
+                    
     img_names = [img.path for img in scenes]
 
-    print('creating image embeddings')
+    print('creating image embeddings...')
 
     img_emb = img_model.encode([Image.open(img.path) for img in scenes], batch_size=128, convert_to_tensor=True, show_progress_bar=True)
 
-    import shutil
-    shutil.rmtree('./scene_outputs')
+    if os.path.exists('./scene_outputs'):
+        import shutil
+        shutil.rmtree('./scene_outputs')
 
     print('searching for prompts...')
-    
+
     for prompt in CLIP_prompts.split(","):
         search(prompt, k=5)
     
@@ -87,10 +85,18 @@ def workflow(video : sieve.Video, font : str, CLIP_prompts : str):
     text_overlay = sieve.function.get('ansh-sievedata-com/image_text_overlay')
     # text_overlay_two = sieve.function.get('ansh-sieve-data/image_text_overlay_two')
 
-    font_path = f'/workflow-custom-fonts/{font}/{font}-Regular.ttf'
+    font_path = f'/root/fonts/workflow-custom-fonts/{font}/{font.replace("_","")}-Regular.ttf'
 
-    base, left, right = random.sample(CLIP_outputs, 3)
-    yield text_overlay.run(base, left, right, video_title, font_path)
+    print('selected font ', font_path)
+
+    for i in range(5):
+        base, left, right = random.sample(CLIP_outputs, 3)
+        yield text_overlay.run(base, left, right, video_title, font_path)
     # yield text_overlay_two.run(base, left, right, video_title, font_path)
 
     print('finished!')
+
+if __name__ == "__main__":
+    print('run?')
+    main(video=sieve.Video(path="./videos/muskcolbert.mp4"), font="Oswald", CLIP_prompts="stephen colbert, elon musk, talkshow")
+    print('done')
